requirements-dev.txt
coverage==5.0.3
django-debug-toolbar==2.1
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
sqlparse==0.3.0

-r requirements.txt
